refactor : main Activity Navigation View 네이밍 수정 완료
