+++ app/src/main/java/com/example/natube/ui/myvideo/MyVideoFragment.kt
@@ -1,4 +1,4 @@
-package com.example.natube.ui.notifications
+package com.example.natube.ui.myvideo
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.natube.databinding.FragmentNotificationsBinding
+import com.example.natube.databinding.FragmentMyVideosBinding
 
-class NotificationsFragment : Fragment() {
+class MyVideoFragment : Fragment() {
 
-    private var _binding: FragmentNotificationsBinding? = null
+    private var _binding: FragmentMyVideosBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class NotificationsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val notificationsViewModel =
-            ViewModelProvider(this).get(NotificationsViewModel::class.java)
+        val myVideoViewModel =
+            ViewModelProvider(this)[MyVideoViewModel::class.java]
 
-        _binding = FragmentNotificationsBinding.inflate(inflater, container, false)
+        _binding = FragmentMyVideosBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
         val textView: TextView = binding.textNotifications
-        notificationsViewModel.text.observe(viewLifecycleOwner) {
+        myVideoViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
