+++ app/src/main/java/com/example/natube/ui/myvideo/MyVideoViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.natube.ui.notifications
+package com.example.natube.ui.myvideo
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class MyVideoViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is notifications Fragment"
