+++ app/src/main/java/com/example/natube/ui/search/SearchViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.natube.ui.dashboard
+package com.example.natube.ui.search
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class SearchViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
